+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Queries.cs
@@ -46,7 +46,7 @@ namespace Stormancer.Server.Plugins.Party
 
         public void Initialize()
         {
-            lucene.TryCreateIndex()
+            lucene.TryCreateIndex(PARTY_LUCENE_INDEX, DefaultMapper.JsonMapper);
         }
     }
 }